python-worker/image_editor_worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0
numpy==1.26.4
torch
torchvision
diffusers
transformers
accelerate
rembg==2.0.59
onnxruntime==1.19.2
opencv-python==4.10.0.84